Remove inher
No more super, no more <
