+++ ulox/ulox.core.tests/Package/Scripts/Samples/06-Class Basics.ulox
@@ -1,12 +1,5 @@
 // classes are defined and then instances can be created.
-class Animal
-{
-	// Classes have methods
-	Speak(){throw;}
-}
-
-// classes can inherit from 1 other class.
-class Cat < Animal
+class Cat
 {
 	Speak(){print("Meow");}
 }
