--- ulox/ulox.core.tests/Package/Scripts/Tests/ClassInherAndInitAndVarInteractions1.ulox
@@ -1,31 +1,0 @@
-class Base
-{
-    init()
-    {
-        this.a = 1;
-    }
-}
-
-class Child < Base
-{
-    init()
-    {
-        this.b = 2;
-    }
-}
-
-test InherClassSuperInitNoParams
-{
-    testcase InherClassSuperInitNoParams
-    {
-        var expected = 4;
-        var result = 0;
-
-        var binst = Base();
-        var cinst = Child();
-
-        result = binst.a + cinst.a + cinst.b;
-
-        Assert.AreEqual(expected, result);
-    }
-}
