--- ulox/ulox.core.tests/Package/Scripts/Tests/ClassInherAndInitAndVarInteractions2.ulox
@@ -1,53 +1,0 @@
-class Base
-{
-    init(a)
-    {
-        this.a = a;
-    }
-}
-
-class Child < Base
-{
-    init(a,b)
-    {
-        this.b = b;
-    }
-}
-
-class Childer < Child
-{
-    init(a,b,c)
-    {
-        this.c = c;
-    }
-}
-
-test InherClassSuperInitParams
-{
-    testcase InherClassSuperInitParams
-    {
-        var expected = 4;
-        var result = 0;
-
-        var binst = Base(1);
-        var cinst = Child(1,2);
-
-        result = binst.a + cinst.a + cinst.b;
-
-        Assert.AreEqual(expected, result);
-    }
-
-    testcase InherClassSuperInitParamsChilder
-    {
-        var expected = 10;
-        var result = 0;
-
-        var binst = Base(1);
-        var cinst = Child(1,2);
-        var cerinst = Childer(1,2,3);
-
-        result = binst.a + cinst.a + cinst.b + cerinst.a + cerinst.b + cerinst.c;
-
-        Assert.AreEqual(expected, result);
-    }
-}
