--- ulox/ulox.core.tests/Package/Scripts/Tests/ClassInherAndInitAndVarInteractions4.ulox
@@ -1,33 +1,0 @@
-class Base
-{
-    var a = 1,b;
-    init()
-    {
-        this.b = 2;
-    }
-}
-
-class Child < Base
-{
-    var c = 3,d;
-    init()
-    {
-        this.d = 4;
-    }
-}
-
-test InherWithVarAndInitNoParams
-{
-    testcase InherWithVarAndInitNoParams
-    {
-        var expected = 13;
-        var result = 0;
-
-        var binst = Base();
-        var cinst = Child();
-
-        result = binst.a + binst.b + cinst.a + cinst.b + cinst.c + cinst.d;
-
-        Assert.AreEqual(expected, result);
-    }
-}
