--- ulox/ulox.core.tests/Package/Scripts/Tests/ClassInherAndInitAndVarInteractions6.ulox
@@ -1,30 +1,0 @@
-class Base
-{
-    var a = 1,b;
-    init(b){}
-}
-
-class Child < Base
-{
-    var c = 3;
-    init(b,d)
-    {
-        this.d = d;
-    }
-}
-
-test InherWithVarAndInitAndAutoVarParams
-{
-    testcase InherWithVarAndInitAndAutoVarParams
-    {
-        var expected = 13;
-        var result = 0;
-
-        var binst = Base(2);
-        var cinst = Child(2,4);
-
-        result = binst.a + binst.b + cinst.a + cinst.b + cinst.c + cinst.d;
-
-        Assert.AreEqual(expected, result);
-    }
-}
