+++ ulox/ulox.core/Package/Runtime/Enums/OpCode.cs
@@ -55,9 +55,6 @@
         METHOD,
         FIELD,
         INVOKE,
-        INHERIT,
-        GET_SUPER,
-        SUPER_INVOKE,
 
         FREEZE,
