+++ ulox/ulox.core/Package/Runtime/Enums/TokenType.cs
@@ -68,7 +68,6 @@
 
         CLASS,
         THIS,
-        SUPER,
         STATIC,
         INIT,
